% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getCP_simon}
\alias{getCP_simon}
\title{Returns the conditional power.}
\usage{
getCP_simon(k, numPat, r1, n1, r, n, p1)
}
\arguments{
\item{k}{number of observed responses}

\item{numPat}{number of enrolled patients.}

\item{r1}{critical value for the first stage.}

\item{n1}{sample size for the first stage.}

\item{r}{critical value for the subset endpoint.}

\item{n}{overall sample size.}

\item{p1}{response rate under the alternative hypothesis.}
}
\description{
Returns the conditional power when "k" responses where observed out of "numPat" Patients for the given Simon's tow stage design.
}
\examples{
#Calculate a Simon's two-stage design
design <- getSolutions()$Solutions[3,] #minimax-design for the default values.
#Assume 3 out of 20 patients had a response.
getCP_simon(3,20,design$r1, design$n1, design$r, design$n, design$p1)
}
