% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setupSimon}
\alias{setupSimon}
\title{Creates a "simon"-object.}
\usage{
setupSimon(alpha = 0.05, beta = 0.05, p0 = 0.1, p1 = 0.3)
}
\arguments{
\item{alpha}{the maximal type I error rate.}

\item{beta}{the maximal type II error rate.}

\item{p0}{the response probability under the null hypothesis.}

\item{p1}{the response probability under the alternative hypothesis.}
}
\description{
Creates a "simon"-object which can be used in the function \code{\link{getSolutions}} to identify possible designs.
}
\examples{
#Create a "simon"-object
simon <- setupSimon()
#Calculate designs for the given "simon"-object.
designs <- getSolutions(simon)$Solutions
designs
}
