\name{Rcpp Modules simon}
\alias{Rcpp_SimonDesign-class}
\alias{SimonDesign}
\title{
  Functions and Objects created by Rcpp for the "Simon's two-stage design"
}
\description{
  The functions and objects of the "simon module" are accessible from R via the
  Rcpp Modules mechanism which creates them based on the declaration
  in the C++ files. The whole implemented functionality of the "simon module"
  is used by the functions implemented in this package. Therefore there is no need
  for the user to access the functions of the module directly.  
}
\references{
  \url{https://CRAN.R-project.org/package=Rcpp}
}
\seealso{
  \code{\link{setupSimon}}, \code{\link{getSolutions}}
}
