\name{Rcpp Modules sub1}
\alias{Rcpp_Sub1Design-class}
\alias{Sub1Design}
\title{
  Functions and Objects created by Rcpp for the "subset design"
}
\description{
  The functions and objects of the "sub1 module" are accessible from R via the
  Rcpp Modules mechanism which creates them based on the declaration
  in the C++ files. The whole implemented functionality of the "sub1 module"
  is used by the functions implemented in this package. Therefore there is no need
  for the user to access the functions of the module directly.
}
\references{
  \url{https://CRAN.R-project.org/package=Rcpp}
}
\seealso{
  \code{\link{setupSub1Design}}, \code{\link{getSolutionsSub1}}
}