% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{toDataframe}
\alias{toDataframe}
\title{Helper function for \code{\link{getSolutions}} and \code{\link{getSolutionsSub1}}.}
\usage{
toDataframe(designObject, useCurtailment = F)
}
\arguments{
\item{designObject}{either a "sub1"-object or a "simon"-object containing the designs for a given parameter set specified inside the "designObject".}

\item{useCurtailment}{boolean value dedicating the use of (non-)stochastic curtailment. This parameter indicates if that information should also gathered.}
}
\description{
Transfers the found designs of a "design"-object to a dataframe. This function is a helper-function used in
\code{\link{getSolutions}} and \code{\link{getSolutionsSub1}} to structure the found designs for a given parameter set specified
in a "sub1"-object or "simon"-object.
The return value is a list containing a dataframe for the found designs and multiple dataframes containing the results for (non-)stochastic curtailment, if present.
}
