% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getP}
\alias{getP}
\title{Calculates the p-value (binomial test).}
\usage{
getP(l, pi0, n2)
}
\arguments{
\item{l}{number of responses}

\item{pi0}{response probability under the null hypothesis}

\item{n2}{number of enrolled patients}
}
\description{
Helper-function for the function \code{\link{getCP}}
}
\seealso{
\code{\link{getCP}}
}
